OPT extract interface for InsertionCostCalculator
